Move G2 Text component to `ui` folder
