+++ packages/components/src/ui/index.js
@@ -1,3 +1,4 @@
 export * from './grid';
+export * from './text';
 export * from './truncate';
 export * from './view';
