+++ packages/components/src/ui/text/index.js
@@ -1,0 +1,2 @@
+export { default as Text } from './text';
+export { default as useText } from './use-text';
