+++ packages/components/src/ui/text/stories/index.js
@@ -1,18 +1,11 @@
 /**
- * External dependencies
- */
-import { boolean } from '@storybook/addon-knobs';
-
-/**
  * Internal dependencies
  */
 import Text from '../text';
-import OldText from '../../index';
-import { adapter } from '../index';
 
 export default {
 	component: Text,
-	title: 'Components/G2/Text',
+	title: 'G2 Components (Experimental)/Text',
 };
 
 export const _default = () => {
@@ -56,46 +49,5 @@ export const highlight = () => {
 			posuere cubilia curae; Vivamus nec pretium velit, sit amet
 			consectetur ante. Praesent porttitor ex eget fermentum mattis.
 		</Text>
-	);
-};
-
-const Adapted = ( { doAdapt, ...props } ) =>
-	doAdapt ? <Text { ...adapter( props ) } /> : <OldText { ...props } />;
-
-export const adapted = () => {
-	const doAdapt = boolean( 'Use adapted component', false );
-	return (
-		<>
-			<Adapted variant="title.large" as="h1" doAdapt={ doAdapt }>
-				Title Large
-			</Adapted>
-			<Adapted variant="title.medium" as="h2" doAdapt={ doAdapt }>
-				Title Medium
-			</Adapted>
-			<Adapted variant="title.small" as="h3" doAdapt={ doAdapt }>
-				Title Small
-			</Adapted>
-			<Adapted variant="subtitle" doAdapt={ doAdapt }>
-				Subtitle
-			</Adapted>
-			<Adapted variant="subtitle.small" doAdapt={ doAdapt }>
-				Subtitle Small
-			</Adapted>
-			<Adapted variant="body" doAdapt={ doAdapt }>
-				Body
-			</Adapted>
-			<Adapted variant="body.small" doAdapt={ doAdapt }>
-				Body Small
-			</Adapted>
-			<Adapted variant="button" doAdapt={ doAdapt }>
-				Button
-			</Adapted>
-			<Adapted variant="caption" doAdapt={ doAdapt }>
-				Caption
-			</Adapted>
-			<Adapted variant="label" doAdapt={ doAdapt }>
-				Label
-			</Adapted>
-		</>
 	);
 };
