+++ packages/components/src/ui/text/text.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { createComponent } from '../../utils';
-import { useText } from './use-text';
+import useText from './use-text';
 
 export default createComponent( {
 	as: 'span',
