+++ packages/components/src/ui/truncate/index.js
@@ -1,3 +1,2 @@
 export { default as Truncate } from './truncate';
-
 export { default as useTruncate } from './use-truncate';
